// src/app/insurances/insurance-type.enum.ts
export enum InsuranceType {
    SCTR = 'SCTR',
    SOAT = 'SOAT',
    VIDALEY = 'VIDALEY',
    POLIZACAR = 'POLIZACAR',
    POLIZATREC = 'POLIZATREC',
    POLIZAEAR = 'POLIZAEAR',
    POLIZAMULTIRRIESGOS = 'POLIZAMULTIRRIESGOS',
    RESCIVIL = 'RESCIVIL',
    VEHICULAR = 'VEHICULAR',
    VIDA = 'VIDA',
    EPS = 'EPS',
    SALUD = 'SALUD',
    ACCIDENTES = 'ACCIDENTES',
    FOLA = 'FOLA',
    VIAJES = 'VIAJES',
    EMPRESARIAL = 'EMPRESARIAL'
}